- Readme; - gulp (minify);
