+++ README.md
@@ -55,4 +55,4 @@ MyApp.controller("MyController", function ($scope, $data, jayTableOptions) {
 
 Take a look at [directive in action](http://gabrielsimplicio.github.io/angular-jaydata-table/) and [demo project](https://github.com/gabrielsimplicio/angular-jaydata-table/tree/master/demo) to see more options.
 
-## Enjoy it! ;)
+#### Enjoy it! ;)
