+++ demo/README.md
@@ -1,0 +1,16 @@
+# angular-jaydata-table-demo
+Demo project for [angular-jaydata-table](https://github.com/gabrielsimplicio/angular-jaydata-table/) directive.
+
+# Installation
+
+#### Instructions
+
+```sh
+$ bower install 
+```
+
+- Open solution at Visual Studio;
+- Open Package Manager Console, and click on "Restore";
+- Start.
+
+#### Enjoy it! ;)
