--- demo/angular-jaydata-table-demo/app/app.js
@@ -1,4 +1,0 @@
-﻿var JayDataDemoApp = angular.module("jayDataDemoApp", [
-    "jaydata",
-    "angular-jaydata-table"
-]);
