+++ demo/angular-jaydata-table-demo/app/app.module.js
@@ -1,0 +1,9 @@
+﻿(function () {
+    'use strict';
+
+    angular.module("jayDataDemoApp", [
+        "jaydata",
+        "angular-jaydata-table"
+    ]);
+
+})();
