+++ demo/angular-jaydata-table-demo/app/customHeaderBody.controller.js
@@ -1,8 +1,11 @@
 ﻿(function () {
     'use strict';
 
-    JayDataDemoApp.controller("ConcatColumnsJavascriptController", function ($scope, $data, jayTableOptions) {
+    var JayDataDemoApp = angular.module("jayDataDemoApp");
 
+    JayDataDemoApp.controller("CustomHeaderBodyController", customHeaderBodyController);
+
+    function customHeaderBodyController($scope, $data) {
         $scope.list = [];
         $scope.selectedItems = [];
 
@@ -11,8 +14,7 @@
             .then(function (odataContext) {
 
                 odataContext
-                    .School
-                    .include("City")
+                    .Person
                     .toArray()
                     .then(function (people) {
                         $scope.$apply(function () {
@@ -23,18 +25,6 @@
             .fail(function (error) {
                 console.log(error);
             });
+    }
 
-        jayTableOptions
-            .initializeHeader()
-            .addColumn("Name")
-            .addColumns()
-                .withTitle("Address/City")
-                .like(function () {
-                    return this.Address + "/" + this.City;
-                })
-                .add("Address").as("Address")
-                .add("City.Name").as("City")
-
-        $scope.options = jayTableOptions;
-    });
 })();
