+++ demo/angular-jaydata-table-demo/app/pagination.controller.js
@@ -1,8 +1,11 @@
 ﻿(function () {
     'use strict';
 
-    JayDataDemoApp.controller("BasicController", function ($scope, $data, jayTableOptions) {
+    var JayDataDemoApp = angular.module("jayDataDemoApp");
 
+    JayDataDemoApp.controller("PaginationController", paginationController);
+
+    function paginationController($scope, $data, jayTableOptions) {
         $scope.list = [];
         $scope.selectedItems = [];
 
@@ -11,13 +14,16 @@
             .then(function (odataContext) {
 
                 odataContext
-                    .School
+                    .Person
+                    .withCount()
+                    .take(10)
                     .toArray()
                     .then(function (people) {
                         $scope.$apply(function () {
                             $scope.list = people;
                         });
                     });
+
             })
             .fail(function (error) {
                 console.log(error);
@@ -26,8 +32,10 @@
         jayTableOptions
             .initializeHeader()
             .addColumn("Name")
-            .addColumn("Address")
+            .addColumn("DateOfBirth").withTitle("Date Of Birth")
+            .addColumn("Phone")
 
         $scope.options = jayTableOptions;
-    });
+    }
+
 })();
