+++ package.json
@@ -2,7 +2,7 @@
   "name": "angular-jaydata-table",
   "version": "1.0.0",
   "description": "AngularJS directive for those using HTML tables and jaydata.",
-  "main": "angular.jaydata.table.js",
+  "main": "dist/angular.jaydata.table.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -20,5 +20,13 @@
   "bugs": {
     "url": "https://github.com/gabrielsimplicio/angular-jaydata-table/issues"
   },
-  "homepage": "https://github.com/gabrielsimplicio/angular-jaydata-table#readme"
+  "homepage": "https://github.com/gabrielsimplicio/angular-jaydata-table#readme",
+  "devDependencies": {
+    "del": "^2.2.0",
+    "gulp": "^3.9.0",
+    "gulp-bower": "0.0.11",
+    "gulp-concat": "^2.6.0",
+    "gulp-ng-annotate": "^1.1.0",
+    "gulp-uglify": "^1.5.1"
+  }
 }
