+++ src/app.module.js
@@ -1,0 +1,5 @@
+(function() {
+    'use strict';
+    
+    angular.module("angular-jaydata-table", ['jaydata']);    
+})();
